+++ packages/sdk/src/create.ts
@@ -1,30 +1,19 @@
-import type { GossipSub } from "@chainsafe/libp2p-gossipsub";
-import { noise } from "@chainsafe/libp2p-noise";
-import { identify } from "@libp2p/identify";
 import type { PeerDiscovery } from "@libp2p/interface";
-import { mplex } from "@libp2p/mplex";
-import { ping } from "@libp2p/ping";
-import { webSockets } from "@libp2p/websockets";
-import { all as filterAll } from "@libp2p/websockets/filters";
-import { wakuFilter, wakuLightPush, wakuMetadata, wakuStore } from "@waku/core";
+import { wakuFilter, wakuLightPush, wakuStore } from "@waku/core";
 import { enrTree, wakuDnsDiscovery } from "@waku/dns-discovery";
 import {
-  type CreateLibp2pOptions,
   DefaultPubsubTopic,
-  type IMetadata,
-  type Libp2p,
   type Libp2pComponents,
   type LightNode,
   type ProtocolCreateOptions,
-  PubsubTopic,
-  type ShardInfo
+  PubsubTopic
 } from "@waku/interfaces";
 import { wakuLocalPeerCacheDiscovery } from "@waku/local-peer-cache-discovery";
 import { wakuPeerExchangeDiscovery } from "@waku/peer-exchange";
 import { ensureShardingConfigured } from "@waku/utils";
-import { createLibp2p } from "libp2p";
 
-import { DefaultUserAgent, WakuNode, WakuOptions } from "./waku.js";
+import { defaultLibp2p } from "./utils/libp2p.js";
+import { WakuNode, WakuOptions } from "./waku.js";
 
 const DEFAULT_NODE_REQUIREMENTS = {
   lightPush: 1,
@@ -130,59 +119,4 @@ export function defaultPeerDiscoveries(
     wakuPeerExchangeDiscovery(pubsubTopics)
   ];
   return discoveries;
-}
-
-type PubsubService = {
-  pubsub?: (components: Libp2pComponents) => GossipSub;
-};
-
-type MetadataService = {
-  metadata?: (components: Libp2pComponents) => IMetadata;
-};
-
-export async function defaultLibp2p(
-  shardInfo?: ShardInfo,
-  wakuGossipSub?: PubsubService["pubsub"],
-  options?: Partial<CreateLibp2pOptions>,
-  userAgent?: string
-): Promise<Libp2p> {
-  if (!options?.hideWebSocketInfo && process.env.NODE_ENV !== "test") {
-    /* eslint-disable no-console */
-    console.info(
-      "%cIgnore WebSocket connection failures",
-      "background: gray; color: white; font-size: x-large"
-    );
-    console.info(
-      "%cWaku tries to discover peers and some of them are expected to fail",
-      "background: gray; color: white; font-size: x-large"
-    );
-    /* eslint-enable no-console */
-  }
-
-  const pubsubService: PubsubService = wakuGossipSub
-    ? { pubsub: wakuGossipSub }
-    : {};
-
-  const metadataService: MetadataService = shardInfo
-    ? { metadata: wakuMetadata(shardInfo) }
-    : {};
-
-  return createLibp2p({
-    connectionManager: {
-      minConnections: 1
-    },
-    transports: [webSockets({ filter: filterAll })],
-    streamMuxers: [mplex()],
-    connectionEncryption: [noise()],
-    ...options,
-    services: {
-      identify: identify({
-        agentVersion: userAgent ?? DefaultUserAgent
-      }),
-      ping: ping(),
-      ...metadataService,
-      ...pubsubService,
-      ...options?.services
-    }
-  }) as any as Libp2p; // TODO: make libp2p include it;
 }
