+++ packages/sdk/src/utils/content_topic.ts
